Moved LoadLanguageFile to testuninstlog.nsi.
When used with MUI2 there was an error saying a language file can't be loaded twice.

In order to be compatible with files that may not use all languages supported by this file it may be necessary to remove the langstring file includes from uninstlog.nsh and require the caller to include the appropriate langstring header files before including uninstlog.nsh.